+++ src/main.ts
@@ -17,22 +17,18 @@ import {
   setupGamePauseOnDialogOpen,
   setupGameShareBtn,
   setupKeyboardShortcuts,
-  setupLevelStartDialog,
   setupPauseBtn,
-  setupResumeBtn,
 } from "./utils/ui/gameplay";
 import { setupScoresDialog } from "./utils/ui/scores";
 import {
   setupContinueGameBtn,
   setupGameOverwriteDialog,
-  setupInstructionsDialog,
   setupNewGameMenuBtn,
 } from "./utils/ui/mainMenu";
 import { setupLoginInvitationDialog } from "./utils/ui/loginInvitationDialog";
 import { setupMusic } from "./utils/audio";
 import { setupMainMenuKeyboardNavigation } from "./utils/ui/mainMenuKeyboardNavigation";
 import { showRandomHint } from "./utils/ui/hints";
-import { setupResetCurrentGameDialog } from "./utils/ui/resetCurrentGameDialog";
 
 document.addEventListener("DOMContentLoaded", async () => {
   registerServiceWorker("cache");
@@ -46,22 +42,18 @@ document.addEventListener("DOMContentLoaded", async () => {
   setupContinueGameBtn();
   setupGameOverwriteDialog();
 
-  setupInstructionsDialog();
   setupLoginButtons();
 
   setupGameControls();
   setupGameShareBtn();
   setupPauseBtn();
-  setupResumeBtn();
   setupKeyboardControls();
   setupCanvasSize();
   setupBackToMenuBtn();
   setupAppVisibilityHandler();
   setupLoginInvitationDialog();
   setupGamePauseOnDialogOpen();
-  setupLevelStartDialog();
   setupKeyboardShortcuts();
-  setupResetCurrentGameDialog();
 
   setupScoresDialog();
   setupFullscreenBtn();