+++ src/utils/ui/customDialog.ts
@@ -4,7 +4,6 @@ import { $id } from "./domQuery";
 export const launchCustomDialog = (title: string, text: string) => {
   const customDialog = $id("customDialog") as PrettyDialog;
   customDialog.open();
-  customDialog.closeButtonIds = ["closeCustomDialogBtn"];
   const customDialogTitle = $id("customDialogTitle");
   if (customDialogTitle) {
     customDialogTitle.innerText = title;