+++ changelog.md
@@ -7,6 +7,8 @@
   - `GameControl`: Drop `callback` setter in favour of `onAction(callback)` method pattern
   - `SocialLink`: Drop `action` setter in favour of `on("click", callback)` method pattern
   - Adapt high scores player check to use the new `guid` field from API 0.5.0
+  - `MenuItem`, `PrettyButton`:
+    `dialog-target` and `dialog-action` attributes replacing `closeButtonIds` array and custom logic
 - Authentication:
   - Implement Facebook SSO
 - Settings: